Fixed time range
For last 3 months was used a formula x - 3. For example if the current month is August the 8th month of the year than 8 - 3 = 5. Thus data shows for May, June, July and August. So for the last 3 month we should subtract 2, for the last 6 months subtract 5, for the last year subtract 11.