Fix wrong denominator values for currencies having digits other than 2 in the database
Fix: display amounts with the minimum currency digits

The migration updates the split values and quantities to match the fractional digits in the currencies.
WARNING: The decimal part will be dropped for currencies that do not use it. This can lead to some data loss if someone was using decimals for a currency which doesn't support it