Execute scheduled actions for all books in the app so all transactions in the books are always up-to-date - fixes #538
Execute scheduled actions service based on the elapsed time since device boot rather than wall clock time. This is more efficient
Change the Service so that the wakelock is always released, regardless of whether the service completed successfully or not - related to #535
Recreate SchedulerService on db upgrade (to schema version13)

Details:
----------
It does not matter if the book is the currently active one or not, it's scheduled transactions will be executed
For this, it is necessary that the SchedulerService use different database adapter instances for each book which it checks.
In turn, it is required to change the db adapters implementation so that none of them use the default adapter instances internally. The default instances always point to the current book, which is not desirable. Every database adapter should only have side-effects on the database on which it was opened.