Improve reliability of scheduled actions - computation of when a scheduled action is due is now based on the execution count (and not the last runtime like before)
Last RunTime now stores the actual time when the scheduled action was executed, which is not necessarily when it was due
Add tests for scheduled actions service
Tapping a book in the BookManager should launch the book

Fix bug causing some scheduled actions to be executed in a loop forever inside the service - fixes #535
There was a bug in the computation of the nextRun time of scheduled actions.
This could lead to situations where the nextRun time remains negative and the service holds a wakelock forever.