Fix widgets with negative values being green
The number used to calculate whether a widget is red or green should be
the same value as the one that is displayed. It was instead calling
a (buggy) method which calculated a different balance to the one
displayed. This patch fixes it to determine the colour using the same
value as the one it is displaying.