Move widget config into separate SharedPreferences
Instead of saving everything into a SharedPreferences object associated
with each book, this saves widget preferences into a separate
SharedPreferences object each.

Some effort is make to ensure that this is backwards compatible, so if
there are settings in the old format they will be copied into the new
format before being erased.