Hide the favorite icon when the item is selected in the account spinner
This is the spinner appearing at the top when listing subaccounts or
transactions. The favorite icon should only be shown when the spinner is
opened. Otherwise it adds clutter and may add confusion with the one
next to it, which is used to mark the account as favorite.

Relates to the previous commit af0c657.