Fix the installation failing with INSTALL_FAILED_CONFLICTING_PROVIDER
The AndroidManifest.xml defines a FileProvider with an authority based
on the reverse domain of the application (org.gnucash.android...). As
it's the same for all build flavors, the installation of two APK of
different flavors fails because the authorities conflict.

This changes the authority to use the application id, which is different
for each flavor.