Ensure default_transfer_account_uid fields don't reference deleted accounts
For some reason we don't have a foreign key constraint on this field, so we
have to enforce it from code. We do so from AccountsDbAdapter.deleteRecord(String)
but not from DatabaseAdapter.deleteRecord(long), which is not overriden
by AccountsDbAdapter (called by AccountsListFragment).

We don't override deleteRecord(long) to avoid infinite recursion (if we call
deleteRecord(String) to avoid code duplication). So we just change
AccountsListFragment to call deleteRecord(String). The proper solution, with a
SQL constraint, will be implemented later when we restructure the database to
make it compatible with the desktop.

Fixes https://github.com/codinguser/gnucash-android/issues/654