Fix export to Google Drive failing with multicurrency QIFs
When transactions with multiple currencies are exported to QIF, a file
for each currency is created in some cases. The problem is the code that
sent the files to Google Drive was creating just one file through
Drive.DriveApi.newDriveContents to which it wrote each of exported
files. For some reason this cause an IOException and nothing was created
on the target directory.

Now we just create a new DriveContents for each exported file.