Fix imported weekly scheduled actions no having the days of the week set
Currently, days of the week aren't parsed for scheduled actions, so they
aren't set either. This ensures scheduled actions are executed at least
once per week by setting them to run on the day of the week of the start
time. It's a temporary workaround, in the future we should implement
parsing of days of the week.