Don't zip exported files when exporting to a URI (SAF)
As with Storage Access Framework we are limited to one file, we checked
if there were multiple files to export and packed them into a ZIP file.
This only happened when exporting to QIF, which now packs multiple files
into a ZIP. Other export formats were already generating a single file.
So now we just copy the exported file.