Don't restrict the files that can be selected on export
When exporting to URI (SAF) we set a MIME type in the intent to select
the target file. By default we used text/*, which restricts the files
allowed to be selected with the file picker to text files. The problem
is XML exports use the .gnca extension, which is not recognized as text.

This changes the MIME type to */* to allow to select any file.