Avoid crashing on start up if no active book is found
This should not happend, but it does. We've received some reports
through Crashlytics.

Nothing obviously wrong has been found. Maybe it got into an invalid
state due to some bug in a previous version. As a consequence, this just
workarounds the problem by setting the first book in the database as
active so the application doesn't crash again.

Fixes:
 http://crashes.to/s/63a0b0221ed
 http://crashes.to/s/740af90dd24
 http://crashes.to/s/e5edc6bf27e