Fix widget black text on black background on newer Android versions
textAppearanceLarge's text color has changed from white to black after
upgrading targetSdkVersion to 26 (56384db). It seems to affect only
newer versions of Android (tested with 7). Switching to
textAppearanceLargeInverted would fix it, but the problem appears on
older versions (tested with 5). So we leave style and just override the
text color.