Fixes slow CSV exports
For each split in every exported transaction, the code fetched the whole
Account instance, including all its transactions with all their splits.
This resulted in hundreds of thousands of requests with a medium-sized
ledger.

There was a cache mechanism, but it was local to one single transaction,
thus essentially useless.

Now, each split generates two requests, to get the account's name and
full name. A better cache mechanism would speed this up further.